Linting update
This is to disable no-member, we were getting a ton of spark_funcs.lit and
spark_funcs.col not being recognized by the linter. I am less concerned about
this, since it is now backed by unit tests. If unit tests exists, and it still
works then this syntax check is not working properly for the situation. One or
the other should exist though.